// types/king.ts
export type King = {
  id: number;
  name: string;
  crest: string;
  government: string;
  counterparts: {
    id: number;
    name: string;
    type: string;
    founder: string;
    location: string;
    primary_gov: string;
    secondary_gov: string;
    economy: string;
    religion: string;
    kingdom_id: number;
  }[];
}